Improve naming convention and fix a bug
* Record has 4 main functions: from_dict, to_dict, persist and select
* select returns a list, entities of which are created via __from_tuple
* fields are stored with prefix '_field_' on their name to sign privacy and prevent beginning with capital letter
* class Decimal integrated to Order and OrderDetail
* something was wrong with handling None in __from_tuple

+++ order.py
@@ -1,10 +1,11 @@
 __author__ = 'KoicsD'
 from record_template import *
+from decimal import Decimal
 
 
 class Order(Record):
     table_name = "Orders"
     fields = (("OrderID", int), ("CustomerID", str), ("EmployeeID", int), ("OrderDate", datetime),
-              ("RequiredDate", datetime), ("ShippedDate", datetime), ("ShipVia", int), ("Freight", float),
+              ("RequiredDate", datetime), ("ShippedDate", datetime), ("ShipVia", int), ("Freight", Decimal),
               ("ShipName", str), ("ShipAddress", str), ("ShipCity", str), ("ShipRegion", str), ("ShipPostalCode", str),
               ("ShipCountry", str))
