+++ order_detail.py
@@ -1,7 +1,8 @@
 __author__ = 'KoicsD'
 from record_template import *
+from decimal import Decimal
 
 
 class OrderDetail(Record):
     table_name = "OrderDetails"
-    fields = (("OrderID", int), ("ProductID", int), ("UnitPrice", float), ("Quantity", int), ("Discount", float))
+    fields = (("OrderID", int), ("ProductID", int), ("UnitPrice", Decimal), ("Quantity", int), ("Discount", float))
